Fix blur issue
When we manually change the date (by typing it) and click "TAB", the first tab will fail.
This is especially annoying when the next field in the form does something obvious when it's focused (such
as show an error). So when the "TAB" fails, it actually TABS and then focuses back which
creates all sorts of problems. This commit allows the first focus out to succeed.